backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic-settings==2.6.1
sqlmodel==0.0.22
psycopg[binary]==3.2.3
alembic==1.13.3
celery==5.4.0
redis==5.1.1
httpx==0.27.2
youtube-transcript-api==1.2.4
python-multipart==0.0.12
tenacity==9.0.0
boto3==1.35.81
pytest==8.3.4
pytest-asyncio==0.24.0
